Refactor service workflows into dedicated modules
Committed-By-Agent: codex
